chore(use case): place a bet use case impl
