+++ app/use-cases/governance/type.ts
@@ -2,6 +2,8 @@ import type {
     CreateGameArgs,
     CreateValidationFunctionArgs
 } from '~/services/governance/type';
+import type { WalletData } from '../wallet';
+import type { PlaceABet } from '../games/type';
 
 export type CreateFunctionData = Omit<
     CreateValidationFunctionArgs,
@@ -9,8 +11,8 @@ export type CreateFunctionData = Omit<
 >;
 export type CreateGameData = Omit<CreateGameArgs, 'signerAddress' | 'provider'>;
 export type PreActionResponse = {
-    wallet?: { address?: string };
-    formData?: CreateFunctionData & CreateGameData;
+    wallet?: WalletData;
+    formData?: CreateFunctionData & CreateGameData & PlaceABet;
     error?: string;
     status?: number;
 };
